+++ ddataflow/data_sources.py
@@ -1,4 +1,4 @@
-from typing import Dict, Any, List
+from typing import Any, Dict, List
 
 from ddataflow.data_source import DataSource